+++ tasks/mpi/koshkin_m_scalar_product_of_vectors/include/ops_mpi.hpp
@@ -15,6 +15,7 @@
 
 namespace koshkin_m_scalar_product_of_vectors {
 int calculateDotProduct(const std::vector<int>& vec_1, const std::vector<int>& vec_2);
+std::vector<int> createRandomVector(int v_size);
 class TestMPITaskSequential : public ppc::core::Task {
  public:
   explicit TestMPITaskSequential(std::shared_ptr<ppc::core::TaskData> taskData_) : Task(std::move(taskData_)) {}
