+++ tasks/mpi/koshkin_m_scalar_product_of_vectors/src/ops_mpi.cpp
@@ -9,6 +9,7 @@
 #include <vector>
 
 using namespace std::chrono_literals;
+static int offset = 0;
 
 bool koshkin_m_scalar_product_of_vectors::TestMPITaskSequential::pre_processing() {
   internal_order_test();
@@ -137,4 +138,12 @@ int koshkin_m_scalar_product_of_vectors::calculateDotProduct(const std::vector<i
   long result = 0;
   for (size_t i = 0; i < vec_1.size(); i++) result += vec_1[i] * vec_2[i];
   return result;
+}
+
+std::vector<int> createRandomVector(int v_size) {
+  std::vector<int> vec(v_size);
+  std::mt19937 gen;
+  gen.seed((unsigned)time(nullptr) + ++offset);
+  for (int i = 0; i < v_size; i++) vec[i] = gen() % 100;
+  return vec;
 }
