+++ tasks/seq/koshkin_m_scalar_product_of_vectors/func_tests/main.cpp
@@ -4,20 +4,11 @@
 #include <random>
 
 #include "seq/koshkin_m_scalar_product_of_vectors/include/ops_seq.hpp"
-static int offset = 0;
-
-std::vector<int> createRandomVector(int v_size) {
-  std::vector<int> vec(v_size);
-  std::mt19937 gen;
-  gen.seed((unsigned)time(nullptr) + ++offset);
-  for (int i = 0; i < v_size; i++) vec[i] = gen() % 100;
-  return vec;
-}
 
 TEST(koshkin_m_scalar_product_of_vectors, check_empty_func_vector_product) {
   const int count = 0;
-  std::vector<int> vec_1 = createRandomVector(count);
-  std::vector<int> vec_2 = createRandomVector(count);
+  std::vector<int> vec_1 = koshkin_m_scalar_product_of_vectors::createRandomVector(count);
+  std::vector<int> vec_2 = koshkin_m_scalar_product_of_vectors::createRandomVector(count);
   int answer = koshkin_m_scalar_product_of_vectors::calculateDotProduct(vec_1, vec_2);
   ASSERT_EQ(0, answer);
 }
@@ -27,8 +18,8 @@ TEST(koshkin_m_scalar_product_of_vectors, check_scalary_product_of_vectors_size_
 
   // Create data
   std::vector<int> out(1, 0);
-  std::vector<int> vec_1 = createRandomVector(count);
-  std::vector<int> vec_2 = createRandomVector(count);
+  std::vector<int> vec_1 = koshkin_m_scalar_product_of_vectors::createRandomVector(count);
+  std::vector<int> vec_2 = koshkin_m_scalar_product_of_vectors::createRandomVector(count);
 
   // Create TaskData
   std::shared_ptr<ppc::core::TaskData> taskDataSeq = std::make_shared<ppc::core::TaskData>();
@@ -57,8 +48,8 @@ TEST(koshkin_m_scalar_product_of_vectors, check_scalary_product_of_vectors_size_
 
   // Create data
   std::vector<int> out(1, 0);
-  std::vector<int> vec_1 = createRandomVector(count);
-  std::vector<int> vec_2 = createRandomVector(count);
+  std::vector<int> vec_1 = koshkin_m_scalar_product_of_vectors::createRandomVector(count);
+  std::vector<int> vec_2 = koshkin_m_scalar_product_of_vectors::createRandomVector(count);
 
   // Create TaskData
   std::shared_ptr<ppc::core::TaskData> taskDataSeq = std::make_shared<ppc::core::TaskData>();
@@ -131,21 +122,14 @@ TEST(koshkin_m_scalar_product_of_vectors, check_calculateDotProduct_binary_neg) 
   ASSERT_EQ(-51, koshkin_m_scalar_product_of_vectors::calculateDotProduct(vec_1, vec_2));
 }
 
-int generateRandomNumber(int min, int max) {
-  std::random_device rd;
-  std::mt19937 gen(rd());
-  std::uniform_int_distribution<> distrib(min, max);
-  return distrib(gen);
-}
-
 TEST(koshkin_m_scalar_product_of_vectors, check_calculateDotProduct_random_size) {
   // Create data
-  int size = generateRandomNumber(1, 100);
+  int size = koshkin_m_scalar_product_of_vectors::generateRandomNumber(1, 100);
   std::vector<int> vec_1(size);
   std::vector<int> vec_2(size);
   for (int i = 0; i < size; ++i) {
-    vec_1[i] = generateRandomNumber(-10, 10);
-    vec_2[i] = generateRandomNumber(-10, 10);
+    vec_1[i] = koshkin_m_scalar_product_of_vectors::generateRandomNumber(-10, 10);
+    vec_2[i] = koshkin_m_scalar_product_of_vectors::generateRandomNumber(-10, 10);
   }
 
   long expected_result = 0;
