+++ tasks/seq/koshkin_m_scalar_product_of_vectors/include/ops_seq.hpp
@@ -8,7 +8,9 @@
 
 namespace koshkin_m_scalar_product_of_vectors {
 int calculateDotProduct(const std::vector<int>& vec_1, const std::vector<int>& vec_2);
-class VectorDotProduct : public ppc::core::Task {
+int generateRandomNumber(int min, int max);
+std::vector<int> createRandomVector(int v_size);
+ class VectorDotProduct : public ppc::core::Task {
  public:
   explicit VectorDotProduct(std::shared_ptr<ppc::core::TaskData> taskData_) : Task(std::move(taskData_)) {}
   bool pre_processing() override;
