+++ tasks/seq/koshkin_m_scalar_product_of_vectors/src/ops_seq.cpp
@@ -1,5 +1,8 @@
 // Copyright 2024 Nesterov Alexander
 #include "seq/koshkin_m_scalar_product_of_vectors/include/ops_seq.hpp"
+#include <random>
+
+static int offset = 0;
 
 bool koshkin_m_scalar_product_of_vectors::VectorDotProduct::pre_processing() {
   internal_order_test();
@@ -44,4 +47,19 @@ int koshkin_m_scalar_product_of_vectors::calculateDotProduct(const std::vector<i
   long result = 0;
   for (size_t i = 0; i < vec_1.size(); i++) result += vec_1[i] * vec_2[i];
   return result;
+}
+
+std::vector<int> koshkin_m_scalar_product_of_vectors::createRandomVector(int v_size) {
+  std::vector<int> vec(v_size);
+  std::mt19937 gen;
+  gen.seed((unsigned)time(nullptr) + ++offset);
+  for (int i = 0; i < v_size; i++) vec[i] = gen() % 100;
+  return vec;
+}
+
+int koshkin_m_scalar_product_of_vectors::generateRandomNumber(int min, int max) {
+  std::random_device rd;
+  std::mt19937 gen(rd());
+  std::uniform_int_distribution<> distrib(min, max);
+  return distrib(gen);
 }
